+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
+- Add tests for the checkboxes in the registration form (#530)
 
 ### Changed
 - Move more tests to the new testing framework (#524, #529)